emergencyresponse_pr/requirements.txt
asgiref==3.11.1
crispy-bootstrap5==2026.3
Django==6.0.2
django-bootstrap5==26.2
django-crispy-forms==2.6
gunicorn==25.1.0
packaging==26.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
